chore(connect): split methods in core
